+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add the ability to navigate by clicking on group members and location occupants on the asset page
 - Add a entity searcher plugin to find entities by their old farmOS 1.x urls
+- Accept pasted images in PhotoInput component
 
 ### Fixed